+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased]
+# 0.14.0 - 2020-10-27
 
 ### Added
 
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated expanding MainTable to use aria-hidden instead of .u-hide utility class
 - Updated Vanilla framework to version 2.19.2
 - Added `positionElementClassName` prop to allow class names to be passed to the position element for the Tooltip component.
+- Bug fixes to Search and Filter component
 
 ### Removed